简化 app.vue
